+++ Classes/IOCTextField.h
@@ -1,2 +1,3 @@
 @interface IOCTextField : UITextField
+@property(nonatomic,assign)CGFloat textRectSubtractOnRight;
 @end